[#50]feat: 폴더 구조 변경 및 for 문의 step이 number에서 string으로 변경돼서 타입 변경
