+++ src/pages/Home/types/conditionItem.ts
