+++ src/pages/Home/types/createToAddData/createToAddForData.ts
@@ -1,0 +1,13 @@
+export interface CreateToAddForData {
+  id: number;
+  depth: number;
+  condition: {
+    target: string;
+    cur: number;
+    start: number;
+    end: number;
+    step: string;
+  };
+  highlights: Array<string>;
+  type: string;
+}
