+++ src/pages/Home/types/dto/variableDto.ts
