+++ src/pages/Home/types/dto/variablesDto.ts
