+++ src/pages/Home/types/printItem.ts
