+++ src/pages/Home/types/variablesItem.ts
