+++ src/pages/Home/types/vizListItem.ts
