+++ src/pages/Home/types/vizVarItem.ts
