--- src/types/createToAddData/createToAddForData.ts
@@ -1,14 +1,0 @@
-export interface CreateToAddForData {
-    id: number;
-    depth: number;
-    condition: {
-      target: string;
-      cur: number;
-      start: number;
-      end: number;
-      step: number;
-    };
-    highlights: Array<string>;
-    type: string;
-  }
-
